Use go 1.24
1. New Cache and CacheWithRenew
2. Move some packages location
