+++ container/map.go
@@ -1,4 +1,4 @@
-package cache
+package container
 
 import "sync"
