+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/sunshineplan/utils
 
-go 1.23
+go 1.24
