+++ loadbalance/loadbalance.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"sync"
 
-	"github.com/sunshineplan/utils/cache"
+	"github.com/sunshineplan/utils/container"
 )
 
 var ErrEmptyLoadBalancer = errors.New("empty load balancer")
@@ -14,7 +14,7 @@ var mu sync.RWMutex
 type LoadBalancer[E any] interface {
 	Len() int
 	Next() E
-	Ring() *cache.Ring[*E]
+	Ring() *container.Ring[*E]
 	Link(LoadBalancer[E]) LoadBalancer[E]
 	Unlink(int) LoadBalancer[E]
 }
