+++ ocr/testdata/ocr.space.logo.png
