+++ smtp/smtp.go
@@ -1,6 +1,7 @@
 package smtp
 
 import (
+	"bytes"
 	"context"
 	"crypto/tls"
 	"encoding/base64"
@@ -213,7 +214,7 @@ func (c *Client) Auth(a smtp.Auth) error {
 	}
 	resp64 := make([]byte, encoding.EncodedLen(len(resp)))
 	encoding.Encode(resp64, resp)
-	code, msg64, err := c.Cmd(0, strings.TrimSpace(fmt.Sprintf("AUTH %s %s", mech, resp64)))
+	code, msg64, err := c.Cmd(0, "AUTH %s %s", mech, bytes.TrimSpace(resp64))
 	for err == nil {
 		var msg []byte
 		switch code {
@@ -239,7 +240,7 @@ func (c *Client) Auth(a smtp.Auth) error {
 		}
 		resp64 = make([]byte, encoding.EncodedLen(len(resp)))
 		encoding.Encode(resp64, resp)
-		code, msg64, err = c.Cmd(0, string(resp64))
+		code, msg64, err = c.Cmd(0, "%s", resp64)
 	}
 	return err
 }
